requirements.txt
gitahttps://github.com/Rapptz/discord.py@rewrite
PyNaCl
pandas
dnspython
async timeout
discord.py
time
random
asyncio
os
